% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{big.xlsx}
\alias{big.xlsx}
\title{Read in a bunch of .xlsx files and spit out a csv}
\usage{
big.xlsx(inpath, out = TRUE, outpath)
}
\arguments{
\item{inpath}{Where is the folder you want to pull the files from? Make sure they have the same columns and there is nothing extra in the folder
out Should the function write out a csv when finished?
outpath If out is TRUE, then where should it be written?}
}
\description{
Read in a bunch of .xlsx files and spit out a csv
}
\examples{
big.xlsx()
}
\keyword{xlsx}

