% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{cor.pairs}
\alias{cor.pairs}
\title{Generate Pairs plot, except instead of giving a scatterplot above and below the main diagonal,
calculate the correlation for each pair and display the correlation either above or below
the main diagonal.}
\usage{
cor.pairs(x, upper = FALSE)
}
\arguments{
\item{x}{dataframe containing variables you want to compare
upper logical defaulted to FALSE. If TRUE, the correlations will be shown in the upper half instead of the lower}
}
\description{
Generate Pairs plot, except instead of giving a scatterplot above and below the main diagonal,
calculate the correlation for each pair and display the correlation either above or below
the main diagonal.
}
\examples{
coolpairs()
}
\keyword{plot}

