% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{directory}
\alias{directory}
\title{Directory}
\usage{
directory()
}
\arguments{
\item{big.csv}{cor.pairs
%not in%
mode
big.xlsx
noNA
what
keyboard}
}
\description{
Directory
}
\examples{
directory()
}
\keyword{directory}

