% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tricks.R
\name{\%not in\%}
\alias{\%not in\%}
\title{Not In}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{the thing you want to filter out
table where you want to filter it out of}
}
\description{
This function serves the opposite function of %in% since R can't figure out !%in%
Found on Stack Overflow. Has the same syntax as %in% (See ?match)
}
\examples{
5 \%not in\% c(1,2,3,4,5)
5 \%not in\% c(1,2,3,4)
}
\keyword{filter}

