% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tricks.R
\name{keyboard}
\alias{keyboard}
\title{Useful RStudio keyboard shortcuts}
\usage{
keyboard()
}
\arguments{
\item{None}{}
}
\description{
Useful RStudio keyboard shortcuts
}
\examples{
keyboard()
}
\keyword{dictionary}

