% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{mode}
\alias{mode}
\title{Mode function: find the mode of a set, functions just like mean or sum functions}
\usage{
mode(x)
}
\arguments{
\item{x}{the list of things you want to find the mode of.}
}
\description{
Mode function: find the mode of a set, functions just like mean or sum functions
}
\examples{
mode()
}
\keyword{stats}

