% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{noNA}
\alias{noNA}
\title{Eliminate NA's}
\usage{
noNA(df, zero = TRUE)
}
\arguments{
\item{df}{The data frame you want to take the NAs out of.}
}
\description{
Eliminate NA's
}
\examples{
noNA(c(1,2,NA,4,5))
noNA(c(1,2,NA,4,5),zero=FALSE)
}
\keyword{NA}

