% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tricks.R
\name{pringle}
\alias{pringle}
\title{Stupid print function that doesn't do anything. I just needed to test adding a function to the package}
\usage{
pringle(string)
}
\arguments{
\item{None}{}
}
\description{
Stupid print function that doesn't do anything. I just needed to test adding a function to the package
}
\examples{
pringle(c("cat",2))
}
\keyword{cat}

