% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tricks.R
\name{reverse}
\alias{reverse}
\title{Reverse character strings or numbers}
\usage{
reverse(text)
}
\arguments{
\item{text}{The text you want to reverse}
}
\description{
I honestly don't know what you'd use this for.
}
\examples{
reverse("Go Hang a Salami, I'm a Lasagna Hog")

reverse(525600)
}
\keyword{character}

