% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{what}
\alias{what}
\title{Reminders of useful code that couldn't easily be written into functions, were already only one function, and shortcuts.}
\usage{
what()
}
\description{
Reminders of useful code that couldn't easily be written into functions, were already only one function, and shortcuts.
}

