% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tricks.R
\name{noNA}
\alias{noNA}
\title{Eliminate NA's}
\usage{
noNA(df, zero = TRUE)
}
\arguments{
\item{df}{The data frame you want to take the NAs out of.}

\item{zero}{Do you want to replace the NAs with '0's? Default set to TRUE. If FALSE, then any row containing an NA will be completely removed.}
}
\description{
Eliminate NA's
}
\examples{
noNA(c(1,2,NA,4,5))
noNA(c(1,2,NA,4,5),zero=FALSE)
}
\keyword{NA}

