% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tricks.R
\name{what}
\alias{what}
\title{Reminders of useful code that couldn't easily be written into functions or were already only one function.}
\usage{
what()
}
\arguments{
\item{None}{just open and close parentheses}
}
\description{
Reminders of useful code that couldn't easily be written into functions or were already only one function.
}
\examples{
what()
}
\keyword{dictionary}

